requirements.txt
asgiref==3.5.2
astroid==2.12.11
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
Django==4.1.1
django-allauth==0.51.0
django-cleanup==6.0.0
django-model-utils==4.2.0
filelock==3.8.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pylint==2.15.4
python-dotenv==0.21.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
singlemodeladmin==0.9
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
termcolor==2.0.1
tomli==2.0.1
tomlkit==0.11.5
virtualenv==20.16.5
wrapt==1.14.1
yapf==0.32.0
